Tanjona comments resolved commit
